# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Journey Orchestrator API
  description: >
    Platform service responsible for executing configurable journeys using a
    state machine model. The service is business-agnostic and manages journey
    definitions and journey instances driven by events.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /journey-instances:
    get:
      tags:
        - Journey Instances
      summary: List all journey instances
      operationId: listJourneyInstances
      responses:
        "200":
          description: List of journey instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JourneyInstanceResponse"
    post:
      tags:
        - Journey Instances
      summary: Start a new journey instance
      operationId: createJourneyInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJourneyInstanceRequest"
      responses:
        "201":
          description: Journey instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyInstanceResponse"
...
  /journey-instances/{instanceId}/connector-executions:
    get:
      tags:
        - Connector Executions
      summary: Get connector execution history for a journey instance
      description: Retrieves the complete connector execution history for a specific journey instance in chronological order
      operationId: getConnectorExecutionHistory
      parameters:
        - name: instanceId
          in: path
          required: true
          description: ID of the journey instance
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Filter history from this timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Filter history to this timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          required: false
          description: Filter by execution status
          schema:
            $ref: "#/components/schemas/ConnectorExecutionStatus"
        - name: limit
          in: query
          required: false
          description: Maximum number of records to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          required: false
          description: Number of records to skip (for pagination)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Connector execution history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorExecutionHistoryResponse"
        '404':
          description: Journey instance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    CreateJourneyInstanceRequest:
      type: object
      required:
        - journeyCode
        - version
      properties:
        journeyCode:
          type: string
        version:
          type: integer
        context:
          type: object
          additionalProperties: true
...
    JourneyInstanceResponse:
      type: object
      properties:
        instanceId:
          type: string
        journeyCode:
          type: string
        version:
          type: integer
        currentState:
          type: string
        status:
          $ref: "#/components/schemas/JourneyStatus"
        context:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JourneyStatus:
      type: string
      enum:
        - RUNNING
        - COMPLETED
        - FAILED
        - CANCELLED
...
    PaginationInfo:
      type: object
      properties:
        limit:
          type: integer
          description: Number of events returned in this page
        offset:
          type: integer
          description: Number of events skipped
        hasNext:
          type: boolean
          description: Whether there are more events available
        hasPrevious:
          type: boolean
          description: Whether there are previous events available
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
    ConnectorType:
      type: string
      enum:
        - HTTP
      description: Type of connector to execute
...
    ConnectorExecutionRecord:
      type: object
      required:
        - id
        - journeyInstanceId
        - stateId
        - connectorType
        - connectorConfiguration
        - startTime
        - endTime
        - duration
        - status
      properties:
        id:
          type: string
          description: Unique identifier for the execution record
        journeyInstanceId:
          type: string
          description: Reference to the journey instance
        stateId:
          type: string
          description: ID of the state that triggered the connector
        connectorType:
          $ref: "#/components/schemas/ConnectorType"
        connectorConfiguration:
          type: string
          description: Serialized configuration (for audit purposes)
        startTime:
          type: string
          format: date-time
          description: Execution start timestamp
        endTime:
          type: string
          format: date-time
          description: Execution end timestamp
        duration:
          type: string
          format: duration
          description: Execution duration (ISO 8601 duration format)
        status:
          $ref: "#/components/schemas/ConnectorExecutionStatus"
        result:
          type: string
          description: Execution result (response data, may be truncated for large payloads)
        errorMessage:
          type: string
          description: Error message if execution failed
    ConnectorExecutionStatus:
      type: string
      enum:
        - SUCCESS
        - FAILURE
    ConnectorExecutionHistoryResponse:
      type: object
      properties:
        instanceId:
          type: string
          description: ID of the journey instance
        executions:
          type: array
          items:
            $ref: "#/components/schemas/ConnectorExecutionRecord"
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
        totalCount:
          type: integer
          description: Total number of execution records